Refactor PTE update code & fix reachable assertion
Deduplicate the code to write the PTE, and in the process remove the reachable (I think) assertion "invalid physical address in TLB".

It actually ended up being basically the same amount of code, so we may want to e.g. inline `update_pte_bits`. However I haven't done that because this code is going to get a bit more complex with CHERI.